Find nearestPair more efficiently.
How this works:

Previously, the nearestPair was found by performing an array.sort() and then
picking off the [0]th element. This is functionally correct, but
computationally wasteful since all the extra sorting work done for the rest
of the array is simply thrown out.

Now we use array.reduce() instead to scan through the array once, looking
for the pair with the smallest distance.